Combined Plough
Currently, various methods are used to apply fertilizer under the plow. At the same time, mineral fertilizers are applied to the soil surface after harvesting in our republic with Amazone ZA M-1500 brand disc centrifugal fertilizer spreaders. Immediately after sowing, the plowing operation is carried out with an SP-12 brand four- body plough so that the applied fertilizers fall under the soil. This is how the operation is performed. The tractor is used twice to carry out this technological operation, which is economically inefficient and inconvenient on slopes. Thus, the centrifugal fertilizer spreader works unevenly. To eliminate this problem, a combined plough was developed in our laboratory. In order to calculate the economic efficiency of the combined plough, the economic indicators of the SP-12 brand four-body plough and the Amazone ZA M- 1500 brand fertilizer spreader were compared, and an economic efficiency of 1586.79 AzN (933.41,dollars $) was obtained. The article provides economic tables. The smooth and even distribution of mineral fertilizers in the field with the application of a combined plough increases the yield in grain growing from 26 cent/ha to 32 cent/ha minimally.
If we calculate the cost of wheat at about 0.35 AzN (0.21 dollars,$) (per kg, then with 600 kg per hectare, it allows to obtain an economic yield of 210 AzN/ ha (123.53 dollars/ha), and an economic yield of 68040 AZN (40,023.53 dollars $) from 324 ha during the season.